requirements.txt
Django~=3.0.5
djangorestframework~=3.11.1
numpy~=1.19.1


face-recognition-models~=0.3.0

pillow~=7.2.0

scikit-learn~=0.23.2
click~=7.1.2
dlib~=19.21.0